feature/317 - Change COLOR_PROFILE enum to THEME
