--- src/assets/scripts/client/constants/colorProfiles.js
@@ -1,94 +1,0 @@
-import { COLORS } from './canvasConstants';
-
-const WIND_VANE_COLOR_DEFAULT = {
-    DIRECTION_LINE: COLORS.WHITE_07,
-    DIRECTION_LINE_GUSTY: COLORS.BRIGHT_RED,
-    INNER_RING_STROKE: COLORS.WHITE_07,
-    OUTER_RING_FILL: COLORS.BLACK_07,
-    WIND_SPEED_TEXT: COLORS.WHITE_07
-};
-
-const DATA_BLOCK_COLOR_IN_RANGE_DEFAULT = {
-    ARRIVAL_BAR: COLORS.RED_05,
-    BACKGROUND: COLORS.GREEN_05,
-    DEPARTURE_BAR: COLORS.BLUE_05,
-    TEXT: COLORS.WHITE_05
-};
-
-const DATA_BLOCK_COLOR_SELECTED_DEFAULT = {
-    ARRIVAL_BAR: COLORS.RED_09,
-    BACKGROUND: COLORS.GREEN_09,
-    DEPARTURE_BAR: COLORS.BLUE_09,
-    TEXT: COLORS.WHITE_09
-};
-
-const DATA_BLOCK_COLOR_OUT_OF_RANGE_DEFAULT = {
-    ARRIVAL_BAR: COLORS.RED_02,
-    BACKGROUND: COLORS.GREEN_02,
-    DEPARTURE_BAR: COLORS.BLUE_02,
-    TEXT: COLORS.WHITE_02
-};
-
-
-const DATA_BLOCK_COLOR_DEFAULT = {
-    IN_RANGE: DATA_BLOCK_COLOR_IN_RANGE_DEFAULT,
-    OUT_OF_RANGE: DATA_BLOCK_COLOR_OUT_OF_RANGE_DEFAULT,
-    SELECTED: DATA_BLOCK_COLOR_SELECTED_DEFAULT
-};
-
-/**
- * @property TERRAIN_COLORS
- * @type {Object}
- * @final
- */
-const TERRAIN_COLOR_DEFAULT = {
-    BORDER_OPACITY: 1,
-    FILL_OPACITY: 0.1,
-    COLORS: {
-        1000: '26, 150, 65',
-        2000: '119, 194, 92',
-        3000: '255, 255, 192',
-        4000: '253, 201, 128',
-        5000: '240, 124, 74',
-        6000: '156, 81, 31'
-    }
-};
-
-// const TERRAIN_COLOR_PROFILES = {
-//     DEFAULT: TERRAIN_COLOR_DEFAULT
-// };
-
-const COLOR_DEFAULT = {
-    AIRSPACE_FILL: COLORS.AIRSPACE_RADIUS_FILL,
-    AIRSPACE_PERIMETER: COLORS.AIRSPACE_RADIUS_STROKE,
-    BACKGROUND: COLORS.MODERATE_GREEN,
-    COMPASS_HASH: COLORS.LIGHT_SILVER,
-    COMPASS_TEXT: COLORS.WHITE,
-    CROSSHAIR_STROKE: COLORS.CROSSHAIR_STROKE,
-    DATA_BLOCK: DATA_BLOCK_COLOR_DEFAULT,
-    FIX_FILL: COLORS.WHITE_05,
-    FIX_TEXT: COLORS.WHITE_05,
-    PROJECTED_TRACK_LINES: COLORS.WHITE,
-    RADAR_TARGET_IN_RANGE: COLORS.WHITE,
-    RADAR_TARGET_OUTSIDE_RANGE: COLORS.LIGHT_SILVER,
-    RADAR_TARGET_PROJECTION_ARRIVAL: COLORS.RED_06,
-    RADAR_TARGET_PROJECTION_DEPARTURE: COLORS.BLUE_06,
-    RADAR_TARGET_PROJECTION_ESTABLISHED_ON_APPROACH: COLORS.RED,
-    RANGE_RING_COLOR: COLORS.AIRSPACE_RANGE_RING_COLOR,
-    RESTRICTED_AIRSPACE: COLORS.RESTRICTED_AIRSPACE_STROKE,
-    RING_CONFLICT: COLORS.WHITE_02,
-    RING_VIOLATION: COLORS.RED,
-    RUNWAY: COLORS.WHITE_04,
-    RUNWAY_EXTENDED_CENTERLINE: COLORS.DARK_PALE_GREEN,
-    RUNWAY_LABELS: COLORS.WHITE_08,
-    SID: COLORS.BLUE_06,
-    TERRAIN: TERRAIN_COLOR_DEFAULT,
-    TOP_ROW_TEXT: COLORS.WHITE_08,
-    TRAILING_SEPARATION_INDICATOR: COLORS.RED_08,
-    WIND_VANE: WIND_VANE_COLOR_DEFAULT,
-    VIDEO_MAP: COLORS.VIDEO_MAP_STROKE
-};
-
-export const COLOR_PROFILES = {
-    DEFAULT: COLOR_DEFAULT
-};
