+++ src/assets/scripts/client/constants/colors/colors.js
@@ -1,0 +1,47 @@
+/**
+ * @property COLORS
+ * @type {Object}
+ * @final
+ */
+export const COLORS = {
+    DARK_PALE_GREEN: 'rgba(68, 102, 85, 1)',
+    MODERATE_GREEN: 'rgba(15, 25, 20, 1)',
+    LIGHT_PALE_GREEN: 'rgba(200, 255, 200, 0.3)',
+    GREEN_02: 'rgba(76, 118, 97, 0.2)',
+    GREEN_05: 'rgba(76, 118, 97, 0.5)',
+    GREEN_09: 'rgba(76, 118, 97, 0.9)',
+    BLACK_07: 'rgba(0, 0, 0, 0.7)',
+    GRAY: 'rgba(0, 0, 0, 0.67)',
+    WHITE_00: 'rgba(255, 255, 255, 0)',
+    WHITE_02: 'rgba(255, 255, 255, 0.2)',
+    WHITE_03: 'rgba(255, 255, 255, 0.2)',
+    WHITE_04: 'rgba(255, 255, 255, 0.4)',
+    WHITE_05: 'rgba(255, 255, 255, 0.5)',
+    WHITE_07: 'rgba(255, 255, 255, 0.7)',
+    WHITE_08: 'rgba(255, 255, 255, 0.8)',
+    WHITE_09: 'rgba(255, 255, 255, 0.9)',
+    WHITE: 'rgba(255, 255, 255, 1.0)',
+    LIGHT_SILVER_03: 'rgba(224, 224, 224, 0.3)',
+    LIGHT_SILVER: 'rgba(224, 224, 224, 1.0)',
+    RED_02: 'rgba(224, 128, 128, 0.2)',
+    RED_05: 'rgba(224, 128, 128, 0.5)',
+    RED_06: 'rgba(224, 128, 128, 0.6)',
+    RED_08: 'rgba(224, 128, 128, 0.8)',
+    RED_09: 'rgba(224, 128, 128, 0.9)',
+    RED: 'rgba(224, 128, 128, 1.0)',
+    CORAL_RED: 'rgba(255, 64, 64, 1.0)',
+    BRIGHT_RED: 'rgba(255, 0, 0, 0.7)',
+    GRAIN_BROWN: 'rgba(224, 210, 180, 1.0)',
+    BLUE_02: 'rgba(128, 255, 255, 0.2)',
+    BLUE_05: 'rgba(128, 255, 255, 0.5)',
+    BLUE_09: 'rgba(128, 255, 255, 0.9)',
+    BLUE_06: 'rgba(128, 255, 255, 0.6)',
+    VIDEO_MAP_STROKE: 'rgba(193, 218, 205, 1)',
+    DEPARTURE_STROKE: 'rgba(128, 255, 255, 0.9)',
+    AIRSPACE_RADIUS_FILL: 'rgba(200, 255, 200, 0.02)',
+    AIRSPACE_RADIUS_STROKE: 'rgba(200, 255, 200, 0.25)',
+    AIRSPACE_RANGE_RING_COLOR: 'rgba(200, 255, 200, 0.1)',
+    CROSSHAIR_STROKE: 'rgba(136, 153, 153, 1)',
+    RESTRICTED_AIRSPACE_STROKE: 'rgba(150, 200, 255, 0.3)',
+    RESTRICTED_AIRSPACE_FILL: 'rgba(150, 200, 255, 0.4)'
+};
