+++ src/assets/scripts/client/constants/colors/dataBlockThemes.js
@@ -1,0 +1,32 @@
+import { COLORS } from './colors';
+
+const DATA_BLOCK_COLOR_IN_RANGE_DEFAULT = {
+    ARRIVAL_BAR: COLORS.RED_05,
+    BACKGROUND: COLORS.GREEN_05,
+    DEPARTURE_BAR: COLORS.BLUE_05,
+    TEXT: COLORS.WHITE_05
+};
+
+const DATA_BLOCK_COLOR_SELECTED_DEFAULT = {
+    ARRIVAL_BAR: COLORS.RED_09,
+    BACKGROUND: COLORS.GREEN_09,
+    DEPARTURE_BAR: COLORS.BLUE_09,
+    TEXT: COLORS.WHITE_09
+};
+
+const DATA_BLOCK_COLOR_OUT_OF_RANGE_DEFAULT = {
+    ARRIVAL_BAR: COLORS.RED_02,
+    BACKGROUND: COLORS.GREEN_02,
+    DEPARTURE_BAR: COLORS.BLUE_02,
+    TEXT: COLORS.WHITE_02
+};
+
+const DATA_BLOCK_COLOR_DEFAULT = {
+    IN_RANGE: DATA_BLOCK_COLOR_IN_RANGE_DEFAULT,
+    OUT_OF_RANGE: DATA_BLOCK_COLOR_OUT_OF_RANGE_DEFAULT,
+    SELECTED: DATA_BLOCK_COLOR_SELECTED_DEFAULT
+};
+
+export const DATA_BLOCK_THEMES = {
+    DEFAULT: DATA_BLOCK_COLOR_DEFAULT
+};
