+++ src/assets/scripts/client/constants/colors/terrainThemes.js
@@ -1,0 +1,21 @@
+/**
+ * @property TERRAIN_COLORS
+ * @type {Object}
+ * @final
+ */
+const TERRAIN_COLOR_DEFAULT = {
+    BORDER_OPACITY: 1,
+    FILL_OPACITY: 0.1,
+    COLORS: {
+        1000: '26, 150, 65',
+        2000: '119, 194, 92',
+        3000: '255, 255, 192',
+        4000: '253, 201, 128',
+        5000: '240, 124, 74',
+        6000: '156, 81, 31'
+    }
+};
+
+export const TERRAIN_THEMES = {
+    DEFAULT: TERRAIN_COLOR_DEFAULT
+};
