+++ src/assets/scripts/client/constants/colors/themes.js
@@ -1,0 +1,39 @@
+import { COLORS } from './colors';
+import { WIND_VANE_THEMES } from './windVaneThemes';
+import { TERRAIN_THEMES } from './terrainThemes';
+import { DATA_BLOCK_THEMES } from './dataBlockThemes';
+
+const DEFAULT = {
+    AIRSPACE_FILL: COLORS.AIRSPACE_RADIUS_FILL,
+    AIRSPACE_PERIMETER: COLORS.AIRSPACE_RADIUS_STROKE,
+    BACKGROUND: COLORS.MODERATE_GREEN,
+    COMPASS_HASH: COLORS.LIGHT_SILVER,
+    COMPASS_TEXT: COLORS.WHITE,
+    CROSSHAIR_STROKE: COLORS.CROSSHAIR_STROKE,
+    DATA_BLOCK: DATA_BLOCK_THEMES.DEFAULT,
+    FIX_FILL: COLORS.WHITE_05,
+    FIX_TEXT: COLORS.WHITE_05,
+    PROJECTED_TRACK_LINES: COLORS.WHITE,
+    RADAR_TARGET_IN_RANGE: COLORS.WHITE,
+    RADAR_TARGET_OUTSIDE_RANGE: COLORS.LIGHT_SILVER,
+    RADAR_TARGET_PROJECTION_ARRIVAL: COLORS.RED_06,
+    RADAR_TARGET_PROJECTION_DEPARTURE: COLORS.BLUE_06,
+    RADAR_TARGET_PROJECTION_ESTABLISHED_ON_APPROACH: COLORS.RED,
+    RANGE_RING_COLOR: COLORS.AIRSPACE_RANGE_RING_COLOR,
+    RESTRICTED_AIRSPACE: COLORS.RESTRICTED_AIRSPACE_STROKE,
+    RING_CONFLICT: COLORS.WHITE_02,
+    RING_VIOLATION: COLORS.RED,
+    RUNWAY: COLORS.WHITE_04,
+    RUNWAY_EXTENDED_CENTERLINE: COLORS.DARK_PALE_GREEN,
+    RUNWAY_LABELS: COLORS.WHITE_08,
+    SID: COLORS.BLUE_06,
+    TERRAIN: TERRAIN_THEMES.DEFAULT,
+    TOP_ROW_TEXT: COLORS.WHITE_08,
+    TRAILING_SEPARATION_INDICATOR: COLORS.RED_08,
+    WIND_VANE: WIND_VANE_THEMES.DEFAULT,
+    VIDEO_MAP: COLORS.VIDEO_MAP_STROKE
+};
+
+export const THEME = {
+    DEFAULT: DEFAULT
+};
