+++ src/assets/scripts/client/constants/colors/windVaneThemes.js
@@ -1,0 +1,13 @@
+import { COLORS } from './colors';
+
+const WIND_VANE_COLOR_DEFAULT = {
+    DIRECTION_LINE: COLORS.WHITE_07,
+    DIRECTION_LINE_GUSTY: COLORS.BRIGHT_RED,
+    INNER_RING_STROKE: COLORS.WHITE_07,
+    OUTER_RING_FILL: COLORS.BLACK_07,
+    WIND_SPEED_TEXT: COLORS.WHITE_07
+};
+
+export const WIND_VANE_THEMES = {
+    DEFAULT: WIND_VANE_COLOR_DEFAULT
+};
